requirements.txt
# Flask和Web应用依赖
Flask==2.2.3
Flask-CORS==3.0.10
Flask-SocketIO==5.3.1
python-socketio==5.7.2
python-engineio==4.3.4
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
eventlet==0.33.3
greenlet==2.0.2

# 网络设备管理
netmiko==4.1.2
paramiko==3.0.0
telnetlib3==1.0.4
pexpect==4.8.0
unipath==1.1

# 数据库
pymysql==1.0.3
mysql-connector-python==8.0.32
SQLAlchemy==2.0.7

# 网络和Web相关
requests==2.28.2
beautifulsoup4==4.11.1
dnspython==2.3.0
websockets==10.4

# PyQt5 GUI
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqtgraph==0.13.3

# 工具和辅助库
six==1.16.0
PyYAML==6.0
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
pyserial==3.5
aiohttp==3.8.4
asyncio==3.4.3

# 日志和监控
loguru==0.6.0
prometheus-client==0.16.0

# 其他工具
python-dotenv==1.0.0
pyinstaller==5.9.0 